requirements.txt
amqp==2.5.1
beautifulsoup4==4.8.1
billiard==3.6.1.0
bs4==0.0.1
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.6
django-celery-results==1.1.2
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.3
django-filter==2.2.0
django-nose==1.4.6
django-pure-pagination==0.3.0
django-rest-framework==0.1.0
django-toolbelt==0.0.1
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
lxml==4.4.1
more-itertools==7.2.0
nose==1.3.7
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
